feat: deprecate inject_richcmp_from_cmp .  Use either GenericRichComparison or total_ordering
GenericRichComparison is specialized against the underlying __attr_comparison__ logic,
barring that, just use functools.total_ordering.

I expect the main (sole?) user of GenericRichComparison will be CPV and atom
classes, since pulling their __cmp__ usage apart will be a PITA.
